minor edit :)

+++ lc_practice/medium/q2.cpp
@@ -18,8 +18,8 @@ struct ListNode {
 class Solution {
 public:
     ListNode* addTwoNumbers(ListNode* l1, ListNode* l2) {
-        long num1 = 0;
-        long num2 = 0;
+        long long num1 = 0;
+        long long num2 = 0;
         while (l1) {
             num1 = num1 * 10 + l1->val;
             l1 = l1->next;
